requirements.txt
mcp==1.6.0
google-genai

pydantic>=2.0
google-generativeai>=0.3.0
python-dotenv>=1.0

rich>=13.0  # for rich console outputs in mcp_server
asyncio
chainlit